src/requirements.txt
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
requests>=2.25.0
